requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0